- Removed methods -- GETDecoder --- GetNumFrames() --- CountFrames() --- SetGraw() --- SetFile() --- IsNextFile()
- Added methods
-- GetDecoder
--- AddGraw()
--- SetData()
--- ShowList()
--- SetNextFile()

Main change is that GETDecoder no longer automatically load the next file
with the numbers at the end of the filename, which is set by CoBo.
This is because the filename from EventBuilder in Narval is different from
the one from CoBo.
GETDecoder now has an array of rawdata files, instead.
Adding files to list won't make autmatically transfer the pointer to the file
if previous file is ended, yet, eventually it will be tranfered.